+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -597,9 +597,9 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
     // MARK: - App Credentials
 
-    func testUsernamePasswordCredential() {
-        let usernamePasswordCredential = Credentials(username: "username", password: "password")
-        XCTAssertEqual(usernamePasswordCredential.provider.rawValue, "local-userpass")
+    func testEmailPasswordCredential() {
+        let EmailPasswordCredential = Credentials(email: "email", password: "password")
+        XCTAssertEqual(EmailPasswordCredential.provider.rawValue, "local-userpass")
     }
 
     func testJWTCredentials() {
@@ -647,12 +647,12 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
     func testInvalidCredentials() {
         do {
-            let username = "testInvalidCredentialsUsername"
+            let email = "testInvalidCredentialsEmail"
             let credentials = basicCredentials()
             let user = try synchronouslyLogInUser(for: credentials)
             XCTAssertEqual(user.state, .loggedIn)
 
-            let credentials2 = Credentials(username: username, password: "NOT_A_VALID_PASSWORD")
+            let credentials2 = Credentials(email: email, password: "NOT_A_VALID_PASSWORD")
             let ex = expectation(description: "Should log in the user properly")
 
             self.app.login(credentials: credentials2, completion: { user2, error in
@@ -737,7 +737,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         let loginEx = expectation(description: "Login user")
         var syncUser: User?
 
-        app.login(credentials: Credentials(username: email, password: password)) { (user, error) in
+        app.login(credentials: Credentials(email: email, password: password)) { (user, error) in
             XCTAssertNil(error)
             syncUser = user
             loginEx.fulfill()
@@ -777,7 +777,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         var syncUser1: User?
         var syncUser2: User?
 
-        app.login(credentials: Credentials(username: email1, password: password1)) { (user, error) in
+        app.login(credentials: Credentials(email: email1, password: password1)) { (user, error) in
             XCTAssertNil(error)
             syncUser1 = user
             login1Ex.fulfill()
@@ -785,7 +785,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
         wait(for: [login1Ex], timeout: 4.0)
 
-        app.login(credentials: Credentials(username: email2, password: password2)) { (user, error) in
+        app.login(credentials: Credentials(email: email2, password: password2)) { (user, error) in
             XCTAssertNil(error)
             syncUser2 = user
             login2Ex.fulfill()
@@ -829,7 +829,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         let loginEx = expectation(description: "Login user")
         var syncUser: User?
 
-        let credentials = Credentials(username: email, password: password)
+        let credentials = Credentials(email: email, password: password)
 
         app.login(credentials: Credentials.anonymous()) { (user, error) in
             XCTAssertNil(error)
@@ -855,7 +855,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
     // MARK: - Provider Clients
 
-    func testUsernamePasswordProviderClient() {
+    func testEmailPasswordProviderClient() {
 
         let email = "realm_tests_do_autoverify\(randomString(7))@\(randomString(7)).com"
         let password = randomString(10)
@@ -924,7 +924,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         wait(for: [registerUserEx], timeout: 4.0)
 
         let loginEx = expectation(description: "Login user")
-        let credentials = Credentials(username: email, password: password)
+        let credentials = Credentials(email: email, password: password)
 
         var syncUser: User?
         app.login(credentials: credentials) { (user, error) in
@@ -999,7 +999,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
         let loginEx = expectation(description: "Login user")
 
-        let credentials = Credentials(username: email, password: password)
+        let credentials = Credentials(email: email, password: password)
         var syncUser: User?
         app.login(credentials: credentials) { (user, error) in
             syncUser = user
@@ -1041,7 +1041,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
 
         let loginExpectation = expectation(description: "Login user")
 
-        let credentials = Credentials(username: email, password: password)
+        let credentials = Credentials(email: email, password: password)
         app.login(credentials: credentials) { (_, error) in
             XCTAssertNil(error)
             loginExpectation.fulfill()
@@ -1077,7 +1077,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         wait(for: [registerUserEx], timeout: 4.0)
 
         let loginEx = expectation(description: "Login user")
-        let credentials = Credentials(username: email, password: password)
+        let credentials = Credentials(email: email, password: password)
         var syncUser: User?
         app.login(credentials: credentials) { (user, error) in
             syncUser = user