+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 x.y.z Release notes (yyyy-MM-dd)
 =============================================================
+### Breaking Changes
+* `Credential(username:password:)` is renamed to `Credential(email:password:)`
+
 ### Enhancements
-* None.
+* Add `User.loggedIn`.
+* Refactor `Credential` and `User` methods for MongoDB Realm consistency.
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-cocoa/issues/????), since v?.?.?)